+++ CHANGELOG.md
@@ -6,6 +6,7 @@
     * Access forms from any other page within the current page
     * Instantiate forms directly in page content with Twig processing enabled
     * New Twig function to get forms data from any other page
+    * Ability to use Twig in saved filename
 
 # v1.3.3
 ## 08/xx/2016