+++ CHANGELOG.md
@@ -7,7 +7,8 @@
     * Instantiate forms directly in page content with Twig processing enabled
     * New Twig function to get forms data from any other page
     * Ability to use Twig in saved filename
-
+1. [](#improved)
+    * Added several missing HTML5 form input field types [#87](https://github.com/getgrav/grav-plugin-form/issues/87)
 # v1.3.3
 ## 08/xx/2016